@@ -3,26 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "~4.6.1",
-    "jquery": "~3.6.0",
+    "bootstrap": "^5.2.0",
+    "jquery": "^3.6.0",
     "merge": "^2.1.1",
     "oidc-client": "~1.11.5",
-    "react": "~16.14.0",
-    "react-dom": "~16.14.0",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router": "~5.2.1",
-    "react-router-dom": "~5.3.0",
-    "reactstrap": "~8.10.1",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.0",
+    "react-dom": "^18.2.0",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "reactstrap": "^9.1.2",
     "rimraf": "^3.0.2"
   },
   "devDependencies": {
     "ajv": "^6.9.1",
     "cross-env": "^5.2.0",
-    "typescript": "^3.7.5",
     "eslint-config-react-app": "^5.2.0",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "nan": "^2.14.1",
-    "react-scripts": "^5.0.0"
+    "react-scripts": "^5.0.0",
+    "typescript": "^3.7.5"
   },
   "eslintConfig": {
     "extends": "react-app"
